Simplify MultiWriteCacheGroup to have only two instances, primary and secondary
